+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `[log]` Added `NilLogger`
 * `[path]` Handle both directions in `DirN`
+* `[strutil]` Code refactoring
 
 ### 12.99.0